Fix tests for the noisy quality report
